fixed compilation failure
